+++ WoodletJump/src/UI/GameOver.cpp
@@ -17,7 +17,7 @@ void GameOver::init(sf::Font &font) {
 
     sf::Text &text = text_.emplace(font, "Game Over!");
     text.setCharacterSize(static_cast<unsigned int>(getUIScale() * kTextDefaultFontSize));
-    float origin = text.getGlobalBounds().size.x / 2.f + text.getLocalBounds().position.x;
+    float origin = text.getGlobalBounds().size.x / 2 + text.getLocalBounds().position.x;
     text.setOrigin({std::round(origin), 0.0f});
     text.setFillColor(sf::Color{180, 50, 50});
 }
@@ -28,7 +28,7 @@ void GameOver::build(Vector2i window_size) {
     int height = static_cast<int>(Constants::kButtonDefaultSize.y * scale);
     int pos_y = window_size.y / 2 - (kTextDefaultFontSize * 2 + height) / 2;
     const int pos_x = window_size.x / 2  - width / 2;
-    text_.value().setPosition({window_size.x / 2.0f, pos_y});   
+    text_.value().setPosition({static_cast<float>(window_size.x / 2), static_cast<float>(pos_y)});   
     pos_y += kTextDefaultFontSize * 2;
     getButton(retry_button_).setBounds({{pos_x, pos_y}, {width, height}});
     UI::build();
